refactor(core): reorg connector routes and UTs
