+++ packages/core/src/routes/connector/index.post.test.ts
@@ -1,7 +1,4 @@
-/* eslint-disable max-lines */
-import type { ConnectorFactory } from '@logto/cli/lib/connector/index.js';
-import { ConnectorPlatform, VerificationCodeType } from '@logto/connector-kit';
-import type { EmailConnector, SmsConnector } from '@logto/connector-kit';
+import { ConnectorPlatform } from '@logto/connector-kit';
 import type { Connector } from '@logto/schemas';
 import { ConnectorType } from '@logto/schemas';
 import { pickDefault, createMockUtils } from '@logto/shared/esm';
@@ -9,13 +6,8 @@ import { any } from 'zod';
 
 import {
   mockMetadata,
-  mockMetadata0,
-  mockMetadata1,
-  mockMetadata2,
-  mockMetadata3,
   mockConnector,
   mockConnectorFactory,
-  mockLogtoConnectorList,
   mockLogtoConnector,
 } from '#src/__mocks__/index.js';
 import RequestError from '#src/errors/RequestError/index.js';
@@ -26,28 +18,14 @@ import type { LogtoConnector } from '#src/utils/connectors/types.js';
 import { createRequester } from '#src/utils/test-utils.js';
 
 const { jest } = import.meta;
-const { mockEsm, mockEsmWithActual } = createMockUtils(jest);
-
-mockEsm('#src/utils/connectors/platform.js', () => ({
-  checkSocialConnectorTargetAndPlatformUniqueness: jest.fn(),
-}));
-
-const removeUnavailableSocialConnectorTargets = jest.fn();
+const { mockEsmWithActual } = createMockUtils(jest);
 
 const connectorQueries = {
-  findConnectorById: jest.fn(),
   countConnectorByConnectorId: jest.fn(),
-  deleteConnectorById: jest.fn(),
   deleteConnectorByIds: jest.fn(),
   insertConnector: jest.fn(async (body) => body as Connector),
 } satisfies Partial<Queries['connectors']>;
-const {
-  findConnectorById,
-  countConnectorByConnectorId,
-  deleteConnectorById,
-  deleteConnectorByIds,
-  insertConnector,
-} = connectorQueries;
+const { countConnectorByConnectorId, deleteConnectorByIds, insertConnector } = connectorQueries;
 
 // eslint-disable-next-line @typescript-eslint/ban-types
 const getLogtoConnectors = jest.fn<Promise<LogtoConnector[]>, []>();
@@ -87,86 +65,13 @@ const tenantContext = new MockTenant(
       return connector;
     },
   },
-  {
-    signInExperiences: { removeUnavailableSocialConnectorTargets },
-  }
+  {}
 );
 
-const connectorRoutes = await pickDefault(import('./connector.js'));
+const connectorDataRoutes = await pickDefault(import('./index.js'));
 
-describe('connector route', () => {
-  const connectorRequest = createRequester({ authedRoutes: connectorRoutes, tenantContext });
-
-  describe('GET /connectors', () => {
-    afterEach(() => {
-      jest.clearAllMocks();
-    });
-
-    it('throws if more than one email connector exists', async () => {
-      getLogtoConnectors.mockResolvedValueOnce(mockLogtoConnectorList);
-      const response = await connectorRequest.get('/connectors').send({});
-      expect(response).toHaveProperty('statusCode', 400);
-    });
-
-    it('throws if more than one SMS connector exists', async () => {
-      getLogtoConnectors.mockResolvedValueOnce(
-        mockLogtoConnectorList.filter((connector) => connector.type !== ConnectorType.Email)
-      );
-      const response = await connectorRequest.get('/connectors').send({});
-      expect(response).toHaveProperty('statusCode', 400);
-    });
-
-    it('shows all connectors', async () => {
-      getLogtoConnectors.mockResolvedValueOnce(
-        mockLogtoConnectorList.filter((connector) => connector.type === ConnectorType.Social)
-      );
-      const response = await connectorRequest.get('/connectors').send({});
-      expect(response).toHaveProperty('statusCode', 200);
-    });
-  });
-
-  describe('GET /connector-factories', () => {
-    it('show all connector factories', async () => {
-      loadConnectorFactories.mockResolvedValueOnce([
-        { ...mockConnectorFactory, metadata: mockMetadata0, type: ConnectorType.Sms },
-        { ...mockConnectorFactory, metadata: mockMetadata1, type: ConnectorType.Social },
-        { ...mockConnectorFactory, metadata: mockMetadata2, type: ConnectorType.Email },
-        { ...mockConnectorFactory, metadata: mockMetadata3, type: ConnectorType.Social },
-      ]);
-      const response = await connectorRequest.get('/connector-factories').send({});
-      expect(response.body).toMatchObject([
-        { ...mockMetadata0, type: ConnectorType.Sms },
-        { ...mockMetadata1, type: ConnectorType.Social },
-        { ...mockMetadata2, type: ConnectorType.Email },
-        { ...mockMetadata3, type: ConnectorType.Social },
-      ]);
-      expect(response).toHaveProperty('statusCode', 200);
-    });
-  });
-
-  describe('GET /connectors/:id', () => {
-    afterEach(() => {
-      jest.clearAllMocks();
-    });
-
-    it('throws when connector can not be found by given connectorId (locally)', async () => {
-      getLogtoConnectors.mockResolvedValueOnce(mockLogtoConnectorList.slice(2));
-      const response = await connectorRequest.get('/connectors/findConnector').send({});
-      expect(response).toHaveProperty('statusCode', 404);
-    });
-
-    it('throws when connector can not be found by given connectorId (remotely)', async () => {
-      getLogtoConnectors.mockResolvedValueOnce([]);
-      const response = await connectorRequest.get('/connectors/id0').send({});
-      expect(response).toHaveProperty('statusCode', 404);
-    });
-
-    it('shows found connector information', async () => {
-      getLogtoConnectors.mockResolvedValueOnce(mockLogtoConnectorList);
-      const response = await connectorRequest.get('/connectors/id0').send({});
-      expect(response).toHaveProperty('statusCode', 200);
-    });
-  });
+describe('connector data route', () => {
+  const connectorRequest = createRequester({ authedRoutes: connectorDataRoutes, tenantContext });
 
   describe('POST /connectors', () => {
     afterEach(() => {
@@ -414,129 +319,4 @@ describe('connector route', () => {
       expect(response).toHaveProperty('statusCode', 422);
     });
   });
-
-  describe('POST /connectors/:id/test', () => {
-    afterEach(() => {
-      jest.clearAllMocks();
-    });
-
-    it('should get SMS connector and send test message', async () => {
-      const sendMessage = jest.fn();
-      const mockedSmsConnectorFactory: ConnectorFactory<SmsConnector> = {
-        ...mockConnectorFactory,
-        metadata: mockMetadata,
-        type: ConnectorType.Sms,
-        createConnector: jest.fn(),
-      };
-      loadConnectorFactories.mockResolvedValueOnce([mockedSmsConnectorFactory]);
-      buildRawConnector.mockResolvedValueOnce({ rawConnector: { sendMessage } });
-      const response = await connectorRequest
-        .post('/connectors/id/test')
-        .send({ phone: '12345678901', config: { test: 123 } });
-      expect(sendMessage).toHaveBeenCalledTimes(1);
-      expect(sendMessage).toHaveBeenCalledWith(
-        {
-          to: '12345678901',
-          type: VerificationCodeType.Generic,
-          payload: {
-            code: '000000',
-          },
-        },
-        { test: 123 }
-      );
-      expect(response).toHaveProperty('statusCode', 204);
-    });
-
-    it('should get email connector and send test message', async () => {
-      const sendMessage = jest.fn();
-      const mockedEmailConnectorFactory: ConnectorFactory<EmailConnector> = {
-        ...mockConnectorFactory,
-        metadata: mockMetadata,
-        type: ConnectorType.Email,
-        createConnector: jest.fn(),
-      };
-      loadConnectorFactories.mockResolvedValueOnce([mockedEmailConnectorFactory]);
-      buildRawConnector.mockResolvedValueOnce({ rawConnector: { sendMessage } });
-      const response = await connectorRequest
-        .post('/connectors/id/test')
-        .send({ email: 'test@email.com', config: { test: 123 } });
-      expect(sendMessage).toHaveBeenCalledTimes(1);
-      expect(sendMessage).toHaveBeenCalledWith(
-        {
-          to: 'test@email.com',
-          type: VerificationCodeType.Generic,
-          payload: {
-            code: '000000',
-          },
-        },
-        { test: 123 }
-      );
-      expect(response).toHaveProperty('statusCode', 204);
-    });
-
-    it('should throw when neither phone nor email is provided', async () => {
-      const response = await connectorRequest.post('/connectors/id/test').send({});
-      expect(response).toHaveProperty('statusCode', 400);
-    });
-
-    it('should throw when sms connector is not found', async () => {
-      getLogtoConnectors.mockResolvedValueOnce([]);
-      const response = await connectorRequest
-        .post('/connectors/id/test')
-        .send({ phone: '12345678901' });
-      expect(response).toHaveProperty('statusCode', 400);
-    });
-
-    it('should throw when email connector is not found', async () => {
-      getLogtoConnectors.mockResolvedValueOnce([]);
-      const response = await connectorRequest
-        .post('/connectors/id/test')
-        .send({ email: 'test@email.com' });
-      expect(response).toHaveProperty('statusCode', 400);
-    });
-  });
-
-  describe('DELETE /connectors/:id', () => {
-    beforeEach(() => {
-      jest.resetAllMocks();
-    });
-
-    afterEach(() => {
-      jest.clearAllMocks();
-    });
-
-    it('delete connector instance and remove unavailable social connector targets', async () => {
-      findConnectorById.mockResolvedValueOnce(mockConnector);
-      loadConnectorFactories.mockResolvedValueOnce([mockConnectorFactory]);
-      await connectorRequest.delete('/connectors/id').send({});
-      expect(deleteConnectorById).toHaveBeenCalledTimes(1);
-      expect(removeUnavailableSocialConnectorTargets).toHaveBeenCalledTimes(1);
-    });
-
-    it('delete connector instance (connector factory is not social type)', async () => {
-      findConnectorById.mockResolvedValueOnce(mockConnector);
-      loadConnectorFactories.mockResolvedValueOnce([
-        { ...mockConnectorFactory, type: ConnectorType.Sms },
-      ]);
-      await connectorRequest.delete('/connectors/id').send({});
-      expect(deleteConnectorById).toHaveBeenCalledTimes(1);
-      expect(removeUnavailableSocialConnectorTargets).toHaveBeenCalledTimes(0);
-    });
-
-    it('delete connector instance (connector factory is not found)', async () => {
-      findConnectorById.mockResolvedValueOnce(mockConnector);
-      loadConnectorFactories.mockResolvedValueOnce([]);
-      await connectorRequest.delete('/connectors/id').send({});
-      expect(deleteConnectorById).toHaveBeenCalledTimes(1);
-      expect(removeUnavailableSocialConnectorTargets).toHaveBeenCalledTimes(0);
-    });
-
-    it('throws when connector not exists with `id`', async () => {
-      // eslint-disable-next-line unicorn/no-useless-undefined
-      findConnectorById.mockResolvedValueOnce(undefined);
-      const response = await connectorRequest.delete('/connectors/id').send({});
-      expect(response).toHaveProperty('statusCode', 500);
-    });
-  });
 });
-/* eslint-enable max-lines */
